Fixed previous PR and id refactor
